Separation of models and services
